# src/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"Encounter": {
				"description": "Encounters are events that can happen to you as a player.\r\nTheir titles must be unique, and a list of actions the player\r\ncan take may be defined.",
				"properties": {
					"title": {
						"type": "string",
						"description": "A fun title for the encounter.",
						"example": "A Zombie Grabs Your Leg!"
					},
					"description": {
						"type": "string",
						"description": "Some text describing what happens in this encounter.",
						"example": "You can tell from here that it has terrible breath."
					},
					"actions": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "A list of actions the user can take when they hit this encounter.",
						"example": [
							"Attempt to kick it away!"
						]
					},
					"numberOfRuns": {
						"type": "number",
						"format": "double",
						"description": "How many times this encounter has been run.\r\nDefaults to 0.",
						"example": 0
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"description": "The date this document was created. Created by Mongoose."
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"description": "The last date this document was updated. Created by Mongoose."
					},
					"__v": {
						"type": "number",
						"format": "double",
						"description": "The version number for this document. Created by Mongoose."
					}
				},
				"required": [
					"title",
					"description",
					"numberOfRuns",
					"createdAt",
					"updatedAt",
					"__v"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseJSON_Encounter-Array_": {
				"properties": {
					"result": {
						"items": {
							"$ref": "#/components/schemas/Encounter"
						},
						"type": "array"
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ErrorResponseJSON": {
				"properties": {
					"error": {
						"type": "string",
						"description": "A description of what went wrong."
					}
				},
				"required": [
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseJSON_Encounter_": {
				"properties": {
					"result": {
						"$ref": "#/components/schemas/Encounter"
					}
				},
				"required": [
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Encounter.title-or-description-or-actions_": {
				"properties": {
					"title": {
						"type": "string",
						"description": "A fun title for the encounter.",
						"example": "A Zombie Grabs Your Leg!"
					},
					"description": {
						"type": "string",
						"description": "Some text describing what happens in this encounter.",
						"example": "You can tell from here that it has terrible breath."
					},
					"actions": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "A list of actions the user can take when they hit this encounter.",
						"example": [
							"Attempt to kick it away!"
						]
					}
				},
				"required": [
					"title",
					"description"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"EncounterCreationParams": {
				"$ref": "#/components/schemas/Pick_Encounter.title-or-description-or-actions_"
			},
			"Partial_EncounterCreationParams_": {
				"properties": {
					"title": {
						"type": "string",
						"description": "A fun title for the encounter.",
						"example": "A Zombie Grabs Your Leg!"
					},
					"description": {
						"type": "string",
						"description": "Some text describing what happens in this encounter.",
						"example": "You can tell from here that it has terrible breath."
					},
					"actions": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "A list of actions the user can take when they hit this encounter.",
						"example": [
							"Attempt to kick it away!"
						]
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"EncounterPatchParams": {
				"$ref": "#/components/schemas/Partial_EncounterCreationParams_"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "Zombie RPG 3000",
		"version": "1.0.0",
		"description": "A server to do all the work behind my zombie roleplaying game",
		"license": {
			"name": "MIT"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/encounters": {
			"get": {
				"operationId": "GetEncounters",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseJSON_Encounter-Array_"
								}
							}
						},
						"description": "Success"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseJSON"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"description": "Get a list of all encounters.\r\nCurrently there is no pagination or limits implemented.",
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "PostEncounter",
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseJSON_Encounter_"
								}
							}
						},
						"description": "Successfully Created"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseJSON"
								}
							}
						},
						"description": "Validation Failed"
					}
				},
				"description": "Creates an encounter.\r\nSupply the title, description, and actions optionally.\r\nThe title must be unique, or else you will receive an error.",
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "A JSON body containing the title, description, and optionally a list of actions.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EncounterCreationParams"
							}
						}
					}
				}
			}
		},
		"/encounters/{id}": {
			"get": {
				"operationId": "GetEncounter",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseJSON_Encounter_"
								}
							}
						},
						"description": "Success"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseJSON"
								}
							}
						},
						"description": "Validation Failed"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseJSON"
								}
							}
						},
						"description": "Encounter Not Found"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseJSON"
								}
							}
						},
						"description": "Internal Server Error"
					}
				},
				"description": "Get a specific encounter by Mongo ID.",
				"security": [],
				"parameters": [
					{
						"description": "Mongo ObjectID of the desired encounter",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "PatchEncounter",
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseJSON_Encounter_"
								}
							}
						},
						"description": "Successfully Updated"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseJSON"
								}
							}
						},
						"description": "Validation Failed"
					},
					"404": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponseJSON"
								}
							}
						},
						"description": "Encounter Not Found"
					}
				},
				"description": "Partially updates an encounter.\r\nProvide only the fields you want to update,\r\nthe rest will remain the same.",
				"security": [],
				"parameters": [
					{
						"description": "Mongo ObjectID of the desired encounter",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Fields and values to update on the encounter",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EncounterPatchParams"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "http://localhost:8000/v1"
		}
	]
}